fix(core,vue): missing subpath exports
