+++ packages/vue/build.config.ts
@@ -9,9 +9,10 @@ export default defineBuildConfig({
   externals: ['vue'],
   entries: [
     { input: 'src/index', name: 'index' },
-    { input: 'src/components/index', name: 'components' },
+    { input: 'src/components', name: 'components' },
     { input: 'src/server', name: 'server' },
     { input: 'src/client', name: 'client' },
     { input: 'src/legacy/index', name: 'legacy' },
+    { input: 'src/types', name: 'types' },
   ],
 })
