+++ packages/vue/src/autoImports.ts
@@ -1,9 +1,5 @@
 import { composableNames } from '@unhead/shared'
 
-const coreComposableNames = [
-  'injectHead',
-]
-
 export const unheadVueComposablesImports = {
-  '@unhead/vue': [...coreComposableNames, ...composableNames],
+  '@unhead/vue': ['injectHead', ...composableNames],
 }
