--- packages/vue/src/components/index.ts
@@ -1,1 +1,0 @@
-export * from './Head'
