--- packages/vue/src/composables/useServerSeoMeta.ts
@@ -1,7 +1,0 @@
-import type { ActiveHeadEntry } from '@unhead/schema'
-import type { UseHeadOptions, UseSeoMetaInput } from '../types'
-import { useSeoMeta } from './useSeoMeta'
-
-export function useServerSeoMeta(input: UseSeoMetaInput, options?: UseHeadOptions): ActiveHeadEntry<UseSeoMetaInput> {
-  return useSeoMeta(input, { ...options, mode: 'server' })
-}
