+++ packages/vue/src/legacy/index.ts
@@ -1,4 +1,4 @@
-import type { ActiveHeadEntry, HeadEntryOptions, MergeHead, Unhead } from '@unhead/schema'
+import type { ActiveHeadEntry, CreateClientHeadOptions, HeadEntryOptions, MergeHead } from '@unhead/schema'
 import type {
   Ref,
 } from 'vue'
@@ -12,6 +12,7 @@ import type {
 } from '../types'
 import { defineHeadPlugin, unpackMeta, whitelistSafeInput } from '@unhead/shared'
 import { createHeadCore } from 'unhead'
+import { DeprecationsPlugin, PromisesPlugin } from 'unhead/plugins'
 import {
   getCurrentInstance,
   inject,
@@ -22,31 +23,56 @@ import {
   watch,
   watchEffect,
 } from 'vue'
-import { createHead } from '../client'
+import { createHead as createVueHead } from '../client'
 import { headSymbol } from '../install'
-import { createHead as createServerHead } from '../server'
+import { createHead as createVueServerHead } from '../server'
 import { resolveUnrefHeadInput } from '../utils'
 
 export * from './useScript'
-export { createHead, createHeadCore, createServerHead, resolveUnrefHeadInput }
+export { createHeadCore, resolveUnrefHeadInput }
 
 export const CapoPlugin = () => defineHeadPlugin({})
+
+export function createHead<T extends MergeHead>(options: CreateClientHeadOptions = {}): VueHeadClient<T> {
+  return createVueHead({
+    disableCapoSorting: true,
+    ...options,
+    plugins: [
+      DeprecationsPlugin,
+      PromisesPlugin,
+      ...(options.plugins || []),
+    ],
+  }) as VueHeadClient<T>
+}
+
+export function createServerHead<T extends MergeHead>(options: CreateClientHeadOptions = {}): VueHeadClient<T> {
+  return createVueServerHead({
+    disableCapoSorting: true,
+    ...options,
+    plugins: [
+      DeprecationsPlugin,
+      PromisesPlugin,
+      ...(options.plugins || []),
+    ],
+  })
+}
 
 /**
  * @deprecated Please switch to non-legacy version
  */
+// eslint-disable-next-line unused-imports/no-unused-vars
 export function setHeadInjectionHandler(handler: () => VueHeadClient<any> | undefined) {
+  // noop
 }
 
 export function injectHead() {
   // fallback to vue context
-  return inject<Unhead>(headSymbol)
+  return inject<VueHeadClient<any>>(headSymbol)
 }
 
 export function useHead<T extends MergeHead>(input: UseHeadInput<T>, options: UseHeadOptions = {}): ActiveHeadEntry<UseHeadInput<T>> | void {
   const head = options.head || injectHead()
   if (head) {
-    // @ts-expect-error untyped
     return head.ssr ? head.push(input, options as HeadEntryOptions) : clientUseHead(head, input, options as HeadEntryOptions)
   }
 }
